fix: Align M2 and M3 Colors
Add ColorScheme to pull colors out of Theme
Add small documentation on how m2 and m3 colors compare

Resolves: 10331
